+++ examples/AspNetCore/Program.cs
@@ -14,8 +14,7 @@
 // limitations under the License.
 // </copyright>
 
-using System.Reflection;
-using OpenTelemetry;
+using Examples.AspNetCore;
 using OpenTelemetry.Exporter;
 using OpenTelemetry.Instrumentation.AspNetCore;
 using OpenTelemetry.Logs;
@@ -37,18 +36,24 @@ var logExporter = appBuilder.Configuration.GetValue<string>("UseLogExporter").To
 // Build a resource configuration action to set service information.
 Action<ResourceBuilder> configureResource = r => r.AddService(
     serviceName: appBuilder.Configuration.GetValue<string>("ServiceName"),
-    serviceVersion: Assembly.GetExecutingAssembly().GetName().Version?.ToString() ?? "unknown",
+    serviceVersion: typeof(Program).Assembly.GetName().Version?.ToString() ?? "unknown",
     serviceInstanceId: Environment.MachineName);
 
+// Create a service to expose ActivitySource, and Metric Instruments
+// for manual instrumentation
+appBuilder.Services.AddSingleton<Instrumentation>();
+
 // Configure OpenTelemetry tracing & metrics with auto-start using the
-// StartWithHost extension from OpenTelemetry.Extensions.Hosting.
+// AddOpenTelemetry extension from OpenTelemetry.Extensions.Hosting.
 appBuilder.Services.AddOpenTelemetry()
     .ConfigureResource(configureResource)
     .WithTracing(builder =>
     {
         // Tracing
 
+        // Ensure the TracerProvider subscribes to any custom ActivitySources.
         builder
+            .AddSource(Instrumentation.ActivitySourceName)
             .SetSampler(new AlwaysOnSampler())
             .AddHttpClientInstrumentation()
             .AddAspNetCoreInstrumentation();
@@ -98,7 +103,9 @@ appBuilder.Services.AddOpenTelemetry()
     {
         // Metrics
 
+        // Ensure the MeterProvider subscribes to any custom Meters.
         builder
+            .AddMeter(Instrumentation.MeterName)
             .AddRuntimeInstrumentation()
             .AddHttpClientInstrumentation()
             .AddAspNetCoreInstrumentation();
@@ -119,8 +126,7 @@ appBuilder.Services.AddOpenTelemetry()
                 builder.AddConsoleExporter();
                 break;
         }
-    })
-    .StartWithHost();
+    });
 
 // Clear default logging providers used by WebApplication host.
 appBuilder.Logging.ClearProviders();