+++ src/OpenTelemetry.Api.ProviderBuilderExtensions/CHANGELOG.md
@@ -2,8 +2,16 @@
 
 ## Unreleased
 
+## 1.4.0-rc.3
+
+Released 2023-Feb-01
+
 * Removed `ConfigureBuilder` from the public API.
   ([#4103](https://github.com/open-telemetry/opentelemetry-dotnet/pull/4103))
+
+* Renamed package from `OpenTelemetry.Extensions.DependencyInjection` to
+  `OpenTelemetry.Api.ProviderBuilderExtensions`.
+  ([#4125](https://github.com/open-telemetry/opentelemetry-dotnet/pull/4125))
 
 ## 1.4.0-rc.2